+++ src/main/java/wooteco/subway/domain/pricing/PricingStrategy.java
@@ -1,6 +1,7 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.pricing;
 
 import java.util.List;
+import wooteco.subway.domain.Section;
 
 public interface PricingStrategy {
