+++ src/main/java/wooteco/subway/service/PathService.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Service;
 import wooteco.subway.dao.SectionDao;
 import wooteco.subway.dao.StationDao;
 import wooteco.subway.domain.Path;
-import wooteco.subway.domain.PathStrategy;
-import wooteco.subway.domain.PricingStrategy;
+import wooteco.subway.domain.path.PathStrategy;
+import wooteco.subway.domain.pricing.PricingStrategy;
 import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Station;
 import wooteco.subway.service.dto.PathResponse;
