+++ src/test/java/wooteco/subway/domain/DistanceProportionalPricingStrategyTest.java
@@ -6,6 +6,8 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import wooteco.subway.domain.pricing.implement.DistanceProportionalPricingStrategy;
+import wooteco.subway.domain.pricing.PricingStrategy;
 
 public class DistanceProportionalPricingStrategyTest {
