+++ src/test/java/wooteco/subway/domain/MinimumDistanceStrategyTest.java
@@ -12,6 +12,8 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
+import wooteco.subway.domain.path.implement.MinimumDistanceStrategy;
+import wooteco.subway.domain.path.PathStrategy;
 
 class MinimumDistanceStrategyTest {
